backend/requirements.txt
Flask
Flask-Cors
Pillow
numpy>=1.20.0,<2.0.0
pathlib>=1.0.1
torch>=2.1.2
torchvision>=0.16.2
lightning>=2.0
omegaconf>=2.1
six>=1.9
einops>=0.4.0 